Added the Time Start and end
Added the code to say when the level started and when it has ended. Used
the get date now function as I felt that the resources for making my own
timer would be better put else where and would slow the game down if i
did put it in as it is the only thing saved is the time stamps and
thouse are only temparary.