Pause and Rank
The pause menu has been implemented so that the camera is disabled.
Problem is that it need to be re-enabled after words. The ranking system
is complete with all ranks and the criteria for the rank is based on how
long it takes to get though the maze. As well the final screen is
skipped by using the space bar